moved comparison ops to `SetAlgebra` extension
